+++ src/api/webrpc/client.go
@@ -55,6 +55,15 @@ func (c *Client) GetUnspentOutputs(addrs []string) (*OutputsResult, error) {
 	return &outputs, nil
 }
 
+func (c *Client) GetUnspentOutputsWithFilters(filters map[string][]string) (*OutputsResult, error) {
+	outputs := OutputsResult{}
+	if err := c.Do(&outputs, "get_outputs_filters", filters); err != nil {
+		return nil, err
+	}
+
+	return &outputs, nil
+}
+
 // InjectTransactionString injects a hex-encoded transaction string to the network
 func (c *Client) InjectTransactionString(rawtx string) (string, error) {
 	params := []string{rawtx}